Refactor: Remove everything tab related
