+++ apps/subscribe-button/lang/de.json
@@ -3,6 +3,7 @@
   "URL": "Feed URL:",
   "COPYURL": "Feed URL kopieren",
   "COPYSUCESS": "URL in die Zwischenablage kopiert.",
+  "PODCATCHER": "Abonnieren mit...",
   "INFO" : {
     "TITLE": "Infos",
     "HEADLINE": "Abonnieren?",
