+++ apps/subscribe-button/src/overlay/components/FinishScreen.vue
