+++ apps/subscribe-button/src/overlay/components/TabHeaderItem.vue
