--- apps/subscribe-button/src/overlay/tabs/index.js
@@ -1,1 +1,0 @@
-export default require('./Tabs').default
