+++ apps/subscribe-button/src/store/reducers/types.js
@@ -1,6 +1,5 @@
 export const INIT = 'INIT'
 export const SET_METADATA = 'SET_METADATA'
-export const TOGGLE_TAB = 'TOGGLE_TAB'
 
 export const SHOW_INFO = 'SHOW_INFO'
 export const HIDE_INFO = 'HIDE_INFO'
