+++ apps/subscribe-button/src/store/selectors.js
@@ -1,14 +1,10 @@
 import * as config from './reducers/config'
 import * as metadata from './reducers/metadata'
 import * as view from './reducers/view'
-import * as tabsreducer from './reducers/tabs'
 import { propOr, compose } from 'ramda'
-
-import { createObject } from '@podlove/utils/helper'
 
 export const configSlice = propOr({}, 'config')
 export const metadataSlice = propOr({}, 'metadata')
-export const tabSlice = propOr({}, 'tabs')
 export const viewSlice = propOr({}, 'view')
 
 export const selectColor = compose(
@@ -71,18 +67,3 @@ export const selectFinishScreenObject = compose(
   view.finish_object,
   viewSlice
 )
-
-export const tabs = createObject({
-  apps: compose(
-    tabsreducer.apps,
-    tabSlice
-  ),
-  web: compose(
-    tabsreducer.web,
-    tabSlice
-  ),
-  info: compose(
-    tabsreducer.info,
-    tabSlice
-  )
-})
